Update default config to use RGB-123-V3 and have 600 pixels per strip
The V3 mapping puts the first 6 channels in a row in order, and they are right near a ground, so easier to understand and test especially for people with less than 7 strings total. 

Increasing the LEDs for strip I think makes a better out-of-the-box experience since it will drive any reasonable length of connected string. Previosuly people would connect long strings and be scared when only the first 64 pixels worked.